Refactoring to split the setup steps over multiple classes and simplify the rule checking
